+++ CHANGELOG.rst
@@ -2,10 +2,15 @@
 1.1.0 (2018-02-XX)
 ==================
 
+- Deconstruct `spoof.HTTPUpstreamServer.handleRequest` to allow more control
+- Add `spoof.HTTPUpstreamServer` to handle proxy requests via CONNECT method
+- Add `spoof.HTTPRequestHandler.do_CONNECT` to handle proxy requests
 - Add `spoof.SSLContext.createOpenSSLConfig` to create self-signed
   certificates with subjectAlternativeName entries, so they can be trusted
   by `requests`, including IP addresses
-- Remove test code that disables SSL warnings
+- Add tests for new functionality
+- Remove test code that disables SSL warnings!
+- Change from nose to pytest for running tests
 - Reformat code to pass Flake8
 - Add `.python-version` pyenv file for testing convenience
 - Re-run tests on latest versions of Python